OSSM-2429 Ensure only specified namespaces are captured (#25)
Before this change, gather_istio.sh captured all the validating/mutating webhook configurations, not just those related to OSSM. When capturing these resources, the associated namespaces are automatically captured by `oc adm inspect`. This caused unexpected namespaces to appear in the must-gather.

In this commit, we ensure that we capture only those validating/mutating webhook configurations that are related to the control plane specified in the `oc adm must-gather` command, thereby ensuring that only the specified namespaces are captured.